check_grid in tests
